# packages/sweep/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sweep API
  description: |
    Multi-chain dust sweeper API that consolidates small token balances across EVM chains and Solana into DeFi yield positions.
    
    ## Authentication
    Most endpoints require authentication via SIWE (Sign-In with Ethereum).
    Send the signature in the `Authorization` header as `Bearer <signature>`.
    
    ## Rate Limits
    - General endpoints: 100 requests/minute
    - Quote endpoints: 50 requests/minute
    - Sweep execution: 10 requests/minute
    
    ## x402 Payments
    Some endpoints require payment via the x402 protocol. When payment is required,
    the API returns HTTP 402 with payment instructions.
  version: 1.0.0
  contact:
    name: Sweep Team
    url: https://github.com/nirholas/sweep
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sweep.exchange
    description: Production
  - url: http://localhost:3000
    description: Development
...
paths:
...
  /api/sweep:
    post:
      tags: [Sweep]
      summary: Execute sweep
      description: |
        Executes a sweep based on a previously generated quote.
        Requires authentication and may require x402 payment ($0.10).
      operationId: executeSweep
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SweepRequest'
      responses:
        '200':
          description: Sweep started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SweepExecuteResponse'
        '402':
          description: Payment required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/X402PaymentRequired'
        '403':
          description: Not authorized
        '404':
          description: Quote not found
        '410':
          description: Quote expired
  /api/sweep/preview:
    post:
      tags: [Sweep]
      summary: Preview sweep
      description: Preview a sweep without executing (no payment required)
      operationId: previewSweep
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet, quoteId]
              properties:
                wallet:
                  type: string
                  pattern: '^0x[a-fA-F0-9]{40}$'
                quoteId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Sweep preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SweepPreviewResponse'
  /api/sweep/{id}/status:
    get:
      tags: [Sweep]
      summary: Get sweep status
      operationId: getSweepStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sweep status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SweepStatusResponse'
        '404':
          description: Sweep not found

  # ============================================================
  # DeFi Endpoints
  # ============================================================
...
  /api/bridge/routes:
    get:
      tags: [Bridge]
      summary: Get bridge routes
      operationId: getBridgeRoutes
      parameters:
        - name: sourceChain
          in: query
          required: true
          schema:
            type: string
            enum: [ethereum, base, arbitrum, polygon, bsc, linea, optimism]
        - name: destChain
          in: query
          required: true
          schema:
            type: string
            enum: [ethereum, base, arbitrum, polygon, bsc, linea, optimism]
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: amount
          in: query
          required: true
          schema:
            type: string
        - name: userAddress
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Available bridge routes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgeRoutesResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: SIWE signature token
  schemas:
...
    SweepRequest:
      type: object
      required: [wallet, quoteId, signature]
      properties:
        wallet:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
        quoteId:
          type: string
          format: uuid
        signature:
          type: string
        gasToken:
          type: string
    SweepExecuteResponse:
      type: object
      properties:
        sweepId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending]
        jobId:
          type: string
        message:
          type: string
        statusUrl:
          type: string
        streamUrl:
          type: string
    SweepPreviewResponse:
      type: object
      properties:
        sweepId:
          type: string
        tokens:
          type: array
          items:
            type: object
        estimatedOutput:
          type: object
        gasEstimate:
          type: object
    SweepStatusResponse:
      type: object
      properties:
        sweepId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        createdAt:
          type: integer
        updatedAt:
          type: integer
        completedAt:
          type: integer
        chains:
          type: array
          items:
            type: string
        tokens:
          type: array
          items:
            type: object
        txHashes:
          type: array
          items:
            type: string
        outputToken:
          type: string
        outputAmount:
          type: string
        outputChain:
          type: string
        totalInputValueUsd:
          type: string
        totalOutputValueUsd:
          type: string
        errorMessage:
          type: string
        defiProtocol:
          type: string
        defiDestination:
          type: string

    # ============================================================
    # x402 Schemas
    # ============================================================
    X402PaymentRequired:
      type: object
      properties:
        x402Version:
          type: integer
          example: 1
        accepts:
          type: array
          items:
            type: object
            properties:
              scheme:
                type: string
                example: exact
              network:
                type: string
                example: base
              asset:
                type: string
                example: '0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913'
              payTo:
                type: string
              maxAmountRequired:
                type: string
                example: '100000'
              resource:
                type: string
              description:
                type: string

    # ============================================================
    # DeFi Schemas
    # ============================================================
...
    BridgeRoutesResponse:
      type: object
      properties:
        success:
          type: boolean
        routes:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              estimatedTime:
                type: integer
              fee:
                type: string
              feeUsd:
                type: number
...